[PATCH] [sailfishos][gecko] Reduce Rust build requirements
Drop debuginfo for Rust components and reduce optimisation when building
swgl.

The style component fails to build on aarch64 when debuginfo is enabled.
The build fails with memory errors and the following:

fatal runtime error: Rust cannot catch foreign exceptions

Disabling debuginfo while building this file allows it to go through.

This change drops debuginfo for all Rust components, so it may be better
to use a more targeted approach instead.

Attempting to build swgl on aarch64 with -O2 level optimisation results
in the error:

internal compiler error: Segmentation fault

Dropping to -O1 allows the build to go through. Updating the compiler
may be a better fix.