[PATCH] [sailfishos][gecko] Restore GLScreenBuffer and TextureImageEGL
TextureImageEGL was removed with Bug 1716559:

https://phabricator.services.mozilla.com/D117904

GLScreenBuffer was removed with Bug 1632249:

https://phabricator.services.mozilla.com/D75055

https://bugzilla.mozilla.org/show_bug.cgi?id=1632249

Both are used as part of the Sailfish OS WebView offscreen render
pipeline. This patch reintroduces these two classes and allows them to
be made use of.

[sailfishos][gecko] Restore GLContextProviderEGL::CreateWrappingExisting()

Restores a method for creating a GLContexst for an existing EglDisplay.
This is useful because on Sailfish OS the EglDisplay is created in
QtMozEmbed and passed to xulrunner, rather than being created by
xulrunner directly.

[sailfishos][gecko] Create GLLibraryEGL as a singleton in GLContextProviderEGL

Ensure that if GLLibraryEGL is creaetd through the GLContextProviderEGL
interface then only one instance is ever created. This ensures that the
internal display management works as expected and avoids a crash that
can occur when a display is freed and the fTerminate method is called.