[PATCH] [sailfishos][gecko] Add and adjust embedlite static prefs
Adds the various embedlite static preferences to StaticPrefList.yaml so
that these will be availble for use as static pref variables in the C++
code.

[sailfishos][gecko] Set security.external_protocol_requires_permission to default to false

Upstream introduces the new user preference:

security.external_protocol_requires_permission

This defaults to true, but we need it to default to false so that
external URL actions continue to work as intended. Here we change the
default value of the preference from true to false so as to allow this.

[sailfishos][gecko] Set embedlite.azpc.json.doubletap static pref to default to true

When the preferences were moved to the StaticPrefList.yaml file the
embedlite.azpc.json.doubletap was set to default to false, but it should
have been set to true. If set to false, double-tap always zooms, even on
items that support double-taps.

[sailfishos][gecko] Set gfx.webrender.force-disabled to default to true

Sailfish OS requires the Webrender pipeline to be disabled. Setting
gfx.webrender.force-disabled on start-up in WebEngineSettings at the app
level isn't enough, so it needs to be patched to default to true in all
cases.