[PATCH] [sailfishos][gecko] Restore NotifyDidPaint event and timers
Reintroduces the NotifyDidPaint event with timers. Ultimately this
ensures that the OnFirstPaint message is sent and received by
EmbedLiteViewParent, so that QMozViewPrivate::OnFirstPaint is called.
Without this, WebView rendering is never initialised on the Qt side.

This is a replacement for the ESR 78 patch that reverts "Bug 1445570 -
Remove EnsureEventualAfterPaint timer", which itself was based on the following
commit:

Revert "Bug 1445570 - Remove EnsureEventualAfterPaint timer. r=tnikkel"

This reverted commit 1deccd7ac14706ad1849343cfb2b93df191a1c42 but the
underlying code has changed significantly so a straightforward revert is
no longer possible.

Co-authored-by: Andrew den Exter <andrew.den.exter@qinetic.com.au>